Rename type_from_filename to file_type
The module is generic enough not to require an overly specific name, and
it exposes more things than simply a default function, so it should be a
noun.

Fixes #6147
