+++ src/cli/plugin/plugin_extractor.js
@@ -1,6 +1,6 @@
 const zipExtract = require('./extractors/zip');
 const tarGzExtract = require('./extractors/tar_gz');
-import { ZIP, TAR } from './type_from_filename';
+import { ZIP, TAR } from './file_type';
 
 export default function extractArchive(settings, logger, archiveType) {
   switch (archiveType) {
